Merge 'stable' into 'main'

+++ Crypter.Contracts/Features/Transfer/DownloadPreview/DownloadTransferFilePreviewResponse.cs
@@ -33,7 +33,7 @@ namespace Crypter.Contracts.Features.Transfer
    {
       public string FileName { get; init; }
       public string ContentType { get; init; }
-      public int Size { get; init; }
+      public long Size { get; init; }
       public string Sender { get; init; }
       public string SenderAlias { get; init; }
       public string Recipient { get; init; }
@@ -43,7 +43,7 @@ namespace Crypter.Contracts.Features.Transfer
       public DateTime ExpirationUTC { get; init; }
 
       [JsonConstructor]
-      public DownloadTransferFilePreviewResponse(string fileName, string contentType, int size, string sender, string senderAlias, string recipient, byte[] publicKey, byte[] keyExchangeNonce, DateTime creationUTC, DateTime expirationUTC)
+      public DownloadTransferFilePreviewResponse(string fileName, string contentType, long size, string sender, string senderAlias, string recipient, byte[] publicKey, byte[] keyExchangeNonce, DateTime creationUTC, DateTime expirationUTC)
       {
          FileName = fileName;
          ContentType = contentType;
